+++ Riot/Modules/Common/SegmentedViewController/SegmentedViewController.m
@@ -135,8 +135,6 @@
     
     // Setup `MXKViewControllerHandling` properties
     self.enableBarTintColorStatusChange = NO;
-    
-    self.sectionHeaderTintColor = ThemeService.shared.theme.tintColor;
 }
 
 - (void)viewDidLoad
@@ -183,6 +181,8 @@
     self.activityIndicator.backgroundColor = ThemeService.shared.theme.overlayBackgroundColor;
     
     self.view.backgroundColor = ThemeService.shared.theme.backgroundColor;
+    
+    self.sectionHeaderTintColor = ThemeService.shared.theme.tintColor;
 }
 
 - (UIStatusBarStyle)preferredStatusBarStyle