+++ Riot/Modules/KeyBackup/Recover/RecoveryKey/KeyBackupRecoverFromRecoveryKeyCoordinator.swift
@@ -66,4 +66,3 @@ extension KeyBackupRecoverFromRecoveryKeyCoordinator: KeyBackupRecoverFromRecove
         self.delegate?.keyBackupRecoverFromPassphraseCoordinatorDidCancel(self)
     }
 }
-