+++ Riot/Modules/KeyBackup/Recover/RecoveryKey/KeyBackupRecoverFromRecoveryKeyViewController.swift
@@ -96,7 +96,7 @@ final class KeyBackupRecoverFromRecoveryKeyViewController: UIViewController {
         
         self.scrollView.keyboardDismissMode = .interactive
         
-        let shieldImage = Asset.Images.shield.image.withRenderingMode(.alwaysTemplate)
+        let shieldImage = Asset.Images.keyBackupLogo.image.withRenderingMode(.alwaysTemplate)
         self.shieldImageView.image = shieldImage
         
         self.informationLabel.text = VectorL10n.keyBackupRecoverFromRecoveryKeyInfo
@@ -161,8 +161,8 @@ final class KeyBackupRecoverFromRecoveryKeyViewController: UIViewController {
         switch viewState {
         case .loading:
             self.renderLoading()
-        case .loaded(totalKeys: let totalKeys):
-            self.renderLoaded(with: totalKeys)
+        case .loaded:
+            self.renderLoaded()
         case .error(let error):
             self.render(error: error)
         }
@@ -173,21 +173,24 @@ final class KeyBackupRecoverFromRecoveryKeyViewController: UIViewController {
         self.activityPresenter.presentActivityIndicator(on: self.view, animated: true)
     }
     
-    private func renderLoaded(with totalKeys: UInt) {
+    private func renderLoaded() {
         self.activityPresenter.removeCurrentActivityIndicator(animated: true)
-        
-        if totalKeys == 0 {
-            self.errorPresenter.presentError(from: self,
-                                             title: VectorL10n.keyBackupRecoverEmptyBackupTitle,
-                                             message: VectorL10n.keyBackupRecoverEmptyBackupMessage,
-                                             animated: true,
-                                             handler: nil)
-        }
     }
     
     private func render(error: Error) {
         self.activityPresenter.removeCurrentActivityIndicator(animated: true)
-        self.errorPresenter.presentError(from: self, forError: error, animated: true, handler: nil)
+
+        if (error as NSError).domain == MXKeyBackupErrorDomain
+            && (error as NSError).code == Int(MXKeyBackupErrorInvalidRecoveryKeyCode.rawValue) {
+
+            self.errorPresenter.presentError(from: self,
+                                             title: VectorL10n.keyBackupRecoverInvalidRecoveryKeyTitle,
+                                             message: VectorL10n.keyBackupRecoverInvalidRecoveryKey,
+                                             animated: true,
+                                             handler: nil)
+        } else {
+            self.errorPresenter.presentError(from: self, forError: error, animated: true, handler: nil)
+        }
     }
     
     private func showFileSelection() {