+++ Riot/Modules/KeyBackup/Setup/Passphrase/KeyBackupSetupPassphraseViewAction.swift
@@ -19,6 +19,7 @@ import Foundation
 /// KeyBackupSetupPassphraseViewController view actions exposed to view model
 enum KeyBackupSetupPassphraseViewAction {
     case setupPassphrase
+    case setupRecoveryKey
     case skip
     case skipAlertSkip
     case skipAlertContinue